[KGP] Add Groovy-Kotlin joint compilation flag
Introduced `groovyKotlinJointCompilation` flag to enable adding `compileKotlin` output to `compileGroovy` classpath. Added tests to validate behavior with and without the flag.

^KT-17153 Fixed
^KT-87200 Fixed
