+++ libraries/tools/kotlin-gradle-plugin/src/common/kotlin/org/jetbrains/kotlin/gradle/plugin/Kotlin2JvmSourceSetProcessor.kt
@@ -10,6 +10,7 @@ import org.gradle.api.tasks.TaskProvider
 import org.gradle.api.tasks.compile.AbstractCompile
 import org.gradle.api.tasks.compile.GroovyCompile
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmCompilerOptions
+import org.jetbrains.kotlin.gradle.dsl.topLevelExtension
 import org.jetbrains.kotlin.gradle.internal.Kapt3GradleSubplugin
 import org.jetbrains.kotlin.gradle.plugin.KotlinPluginLifecycle.Stage.AfterEvaluateBuildscript
 import org.jetbrains.kotlin.gradle.scripting.internal.ScriptingGradleSubplugin
@@ -63,11 +64,13 @@ internal class Kotlin2JvmSourceSetProcessor(
                 javaTask.configure { javaCompile ->
                     javaCompile.classpath += project.files(kotlinTask.flatMap { it.destinationDirectory })
                 }
-                project.tasks.withType<GroovyCompile>()
-                    .configureEach { groovyCompile ->
-                        if (groovyCompile.name != java.getCompileTaskName("groovy")) return@configureEach
-                        groovyCompile.classpath += project.files(kotlinTask.map { it.destinationDirectory.get() })
-                    }
+                if (project.topLevelExtension.groovyKotlinJointCompilation) {
+                    project.tasks.withType<GroovyCompile>()
+                        .configureEach { groovyCompile ->
+                            if (groovyCompile.name != java.getCompileTaskName("groovy")) return@configureEach
+                            groovyCompile.classpath += project.files(kotlinTask.map { it.destinationDirectory.get() })
+                        }
+                }
                 kotlinTask.configure { kotlinCompile ->
                     kotlinCompile.javaOutputDir.set(javaTask.flatMap { it.destinationDirectory })
                 }
